Moved models.UserRoleName to database.models.UserRoleName
When Streamlit Passwordless is initialized the default roles are created
in the database. It makes more sense to keep the `UserRoleName` enum among
the database models. Updated the methods `create_viewer`, `create_user`,
`create_superuser` and `create_admin` of `models.Role` to use their counterpart
methods from `database.models.Role`, which avoids duplication of these methods.
