+++ src/streamlit_passwordless/database/models.py
@@ -3,6 +3,7 @@ r"""The models that represent tables in the database."""
 # Standard library
 import os
 from datetime import datetime
+from enum import StrEnum
 from typing import ClassVar, Optional, Self
 
 # Third party
@@ -14,9 +15,6 @@ from sqlalchemy.orm import (
     mapped_column,
     relationship,
 )
-
-# Local
-from streamlit_passwordless.models import UserRoleName
 
 SCHEMA: str | None = os.getenv('STP_DB_SCHEMA')
 metadata_obj = MetaData(schema=SCHEMA)
@@ -97,6 +95,36 @@ class Base(DeclarativeBase):
         output += f'{self._modified_at_created_at_as_str}\n)'
 
         return output
+
+
+class UserRoleName(StrEnum):
+    r"""The predefined user role names of streamlit-passwordless.
+
+    These roles are created in the database when the database is initialized.
+    The default role of a new user is :attr:`UserRoleName.USER`.
+
+    Members
+    -------
+    VIEWER
+        A user that can only view data within an application.
+
+    USER
+        The standard user with normal privileges. When a user is created it is
+        assigned this role by default.
+
+    SUPERUSER
+        A user with higher privileges that can perform certain
+        operations that a normal `USER` can not.
+
+    ADMIN
+        An admin has full access to everything. Only admin users may sign in to the admin page
+        and manage the users of the application. An application should have at least one admin.
+    """
+
+    VIEWER = 'Viewer'
+    USER = 'User'
+    SUPERUSER = 'SuperUser'
+    ADMIN = 'Admin'
 
 
 class Role(Base):
